requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-allauth==0.44.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
html5lib==1.0.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.3
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==8.1.2
pkg-resources==0.0.0
progress==1.5
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.14.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.4.2
webencodings==0.5.1
